+++ packages/calcite-components/src/tests/commonTests/t9n.ts
@@ -3,7 +3,7 @@ import { E2EElement, E2EPage } from "@stencil/core/testing";
 import { toHaveNoViolations } from "jest-axe";
 import { MessageBundle } from "../../utils/t9n";
 import { IntrinsicElementsWithProp } from "./../utils";
-import { getTagAndPage, ComponentTestSetup } from "./setupForTests";
+import { getTagAndPage, ComponentTestSetup } from "./utils";
 
 expect.extend(toHaveNoViolations);
